refactor: entity post 슬라이스 관련 리팩토링 진행
