+++ apps/web/src/entities/post/config/index.ts
@@ -1,1 +1,2 @@
-export * from './query-key';
+export { POST_API_PREFIX } from './postApiPrefix';
+export { postQueryKeys, postQueryOptions } from './query';
