+++ apps/web/src/entities/post/config/postApiPrefix.ts
@@ -1,0 +1,1 @@
+export const POST_API_PREFIX = '/api/v2/posts';
