--- apps/web/src/entities/post/config/query-key/index.ts
@@ -1,5 +1,0 @@
-export const postKeys = {
-  all: ['post'] as const,
-
-  detail: (postId: string) => [...postKeys.all, postId] as const,
-};
