+++ apps/web/src/entities/post/config/query/index.ts
@@ -1,0 +1,2 @@
+export { postQueryKeys } from './postQueryKeys';
+export { postQueryOptions } from './postQueryOptions';
