+++ apps/web/src/entities/post/config/query/postQueryKeys.ts
@@ -1,0 +1,4 @@
+export const postQueryKeys = {
+  all: ['post'] as const,
+  detail: (postId: string) => [...postQueryKeys.all, postId] as const,
+};
