+++ apps/web/src/entities/post/model/types/dto/getPostDto.ts
@@ -1,0 +1,23 @@
+/* 게시글 단일 조회 */
+export interface GetPostResponseDto {
+  id: string;
+  content: string;
+  isDeleted: boolean;
+  displayWriterNickname: string;
+  writerProfileImage: string;
+  fileUrlList: string[];
+  numComment: number;
+  numLike: number;
+  numFavorite: number;
+  voteId?: string;
+  isAnonymous: boolean;
+  isOwner: boolean;
+  isPostLike: boolean;
+  isPostFavorite: boolean;
+  updatable: boolean;
+  deletable: boolean;
+  createdAt: string;
+  updatedAt: string;
+  boardId: string;
+  boardName: string;
+}
