+++ apps/web/src/entities/post/model/types/dto/index.ts
@@ -1,0 +1,1 @@
+export { type GetPostResponseDto } from './getPostDto';
