+++ apps/web/src/entities/post/model/types/index.ts
@@ -1,3 +1,4 @@
 export * from './detail';
 export * from './write';
 export * from './update';
+export * from './dto';
